refactor: separate DTOs into request and response packages
