+++ src/main/java/com/muud/auth/controller/AuthController.java
@@ -1,6 +1,11 @@
 package com.muud.auth.controller;
 
-import com.muud.auth.domain.dto.*;
+import com.muud.auth.domain.dto.request.SigninRequest;
+import com.muud.auth.domain.dto.request.SignupRequest;
+import com.muud.auth.domain.dto.response.KakaoInfoResponse;
+import com.muud.auth.domain.dto.response.SigninResponse;
+import com.muud.auth.domain.dto.response.SignupResponse;
+import com.muud.auth.domain.dto.response.TokenResponse;
 import com.muud.auth.service.AuthService;
 import com.muud.auth.service.KakaoService;
 import com.muud.global.error.ApiException;
