+++ src/main/java/com/muud/auth/domain/dto/request/RefreshTokenRequest.java
@@ -1,2 +1,4 @@
-package com.muud.auth.domain.dto.request;public class RefreshTokenRequest {
+package com.muud.auth.domain.dto.request;
+
+public class RefreshTokenRequest {
 }
