+++ src/main/java/com/muud/auth/domain/dto/response/TokenResponse.java
@@ -1,4 +1,4 @@
-package com.muud.auth.domain.dto;
+package com.muud.auth.domain.dto.response;
 
 public record TokenResponse(String accessToken) {
