+++ src/main/java/com/muud/auth/service/AuthService.java
@@ -1,6 +1,10 @@
 package com.muud.auth.service;
 
-import com.muud.auth.domain.dto.*;
+import com.muud.auth.domain.dto.request.SigninRequest;
+import com.muud.auth.domain.dto.request.SignupRequest;
+import com.muud.auth.domain.dto.response.KakaoInfoResponse;
+import com.muud.auth.domain.dto.response.SigninResponse;
+import com.muud.auth.domain.dto.response.TokenResponse;
 import com.muud.auth.exception.AuthException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
