+++ src/test/java/com/muud/auth/service/AuthServiceTest.java
@@ -1,6 +1,9 @@
 package com.muud.auth.service;
 
-import com.muud.auth.domain.dto.*;
+import com.muud.auth.domain.dto.request.SigninRequest;
+import com.muud.auth.domain.dto.request.SignupRequest;
+import com.muud.auth.domain.dto.response.KakaoInfoResponse;
+import com.muud.auth.domain.dto.response.SigninResponse;
 import com.muud.auth.exception.AuthErrorCode;
 import com.muud.auth.exception.AuthException;
 import com.muud.auth.jwt.JwtToken;
